requirements.txt
absl-py==2.2.1
acres==0.3.0
altair==5.5.0
altgraph==0.17.4
annotated-types==0.7.0
astunparse==1.6.3
attrs==25.3.0
beautifulsoup4==4.13.3
blinker==1.9.0
blis==1.3.0
cachetools==5.5.2
catalogue==2.0.10
certifi==2025.1.31
charset-normalizer==3.4.1
ci-info==0.3.0
click==8.1.8
cloudpathlib==0.21.0
colorama==0.4.6
confection==0.1.5
configobj==5.0.9
configparser==7.2.0
cymem==2.0.11
Deprecated==1.2.18
en_core_web_sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.8.0/en_core_web_sm-3.8.0-py3-none-any.whl#sha256=1932429db727d4bff3deed6b34cfc05df17794f4a52eeb26cf8928f7c1a0fb85
et_xmlfile==2.0.0
etelemetry==0.3.1
Faker==37.1.0
filelock==3.18.0
fitz==0.0.1.dev2
Flask==3.1.0
Flask-Limiter==3.12
flatbuffers==25.2.10
fsspec==2025.3.2
gast==0.6.0
gitdb==4.0.12
GitPython==3.1.44
google-pasta==0.2.0
grpcio==1.71.0
h5py==3.13.0
httplib2==0.22.0
huggingface-hub==0.30.1
idna==3.10
importlib_resources==6.5.2
isodate==0.6.1
itsdangerous==2.2.0
Jinja2==3.1.6
joblib==1.4.2
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
keras==3.9.1
langcodes==3.5.0
language_data==1.3.0
libclang==18.1.1
limits==4.6
looseversion==1.3.0
lxml==5.3.1
marisa-trie==1.2.1
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
ml_dtypes==0.5.1
mpmath==1.3.0
murmurhash==1.0.12
namex==0.0.8
narwhals==1.34.0
networkx==3.4.2
nibabel==5.3.2
nipype==1.10.0
numpy==2.1.3
openpyxl==3.1.5
opt_einsum==3.4.0
optree==0.14.1
ordered-set==4.1.0
packaging==24.2
pandas==2.2.3
pathlib==1.0.1
pefile==2023.2.7
pillow==11.1.0
preshed==3.0.9
protobuf==5.29.4
prov==2.0.1
puremagic==1.28
pyarrow==19.0.1
pydantic==2.11.2
pydantic_core==2.33.1
pydeck==0.9.1
pydot==3.0.4
Pygments==2.19.1
pyinstaller==6.12.0
pyinstaller-hooks-contrib==2025.2
PyMuPDF==1.25.5
pyparsing==3.2.3
PyPDF2==3.0.1
python-dateutil==2.9.0.post0
python-docx==1.1.2
pytz==2025.2
pywin32-ctypes==0.2.3
pyxnat==1.6.3
PyYAML==6.0.2
rdflib==6.3.2
referencing==0.36.2
regex==2024.11.6
requests==2.32.3
rich==13.9.4
rpds-py==0.24.0
safetensors==0.5.3
scikit-learn==1.6.1
scipy==1.15.2
shellingham==1.5.4
simplejson==3.20.1
six==1.17.0
smart-open==7.1.0
smmap==5.0.2
soupsieve==2.6
spacy==3.8.5
spacy-legacy==3.0.12
spacy-loggers==1.0.5
srsly==2.5.1
streamlit==1.44.1
sympy==1.13.1
tenacity==9.1.2
tensorboard==2.19.0
tensorboard-data-server==0.7.2
tensorflow-io-gcs-filesystem==0.31.0
termcolor==3.0.0
thinc==8.3.6
threadpoolctl==3.6.0
tokenizers==0.21.1
toml==0.10.2
torch==2.6.0
tornado==6.4.2
tqdm==4.67.1
traits==7.0.2
transformers==4.50.3
typer==0.15.2
typing-inspection==0.4.0
typing_extensions==4.12.2
tzdata==2025.2
ua-parser==1.0.1
ua-parser-builtins==0.18.0.post1
urllib3==2.3.0
user-agents==2.2.0
wasabi==1.1.3
watchdog==6.0.0
weasel==0.4.1
Werkzeug==3.1.3
wrapt==1.17.2
